Updated Coffee break download.
Now coffee break is shared in the RSS with two mp3 for each
chapter, side A and B. So we download them based on the title
checking for a pattern and the we join both files in one.

We are also using this for other radios using the download
system.